+++ lib/modules/platform/index.ts
@@ -79,26 +79,5 @@ export async function initPlatform(config: AllConfig): Promise<AllConfig> {
   };
   returnConfig.hostRules.push(typedPlatformRule);
   hostRules.add(typedPlatformRule);
-  const gitHubRule: HostRule[] = [
-    {
-      matchHost: '.pkg.github.com',
-    },
-    {
-      matchHost: 'ghcr.io',
-    },
-  ];
-  for (const rule of gitHubRule) {
-    (
-      ['token', 'username', 'password'] as ('token' | 'username' | 'password')[]
-    ).forEach((field) => {
-      if (config[field]) {
-        rule[field] = config[field] as string;
-        delete returnConfig[field];
-      }
-    });
-    returnConfig.hostRules = returnConfig.hostRules || [];
-    returnConfig.hostRules.push(rule);
-    hostRules.add(rule);
-  }
   return returnConfig;
 }
